feat:Added monogodb

+++ Backend/src/controllers/address.ts
@@ -1,12 +1,17 @@
 import { RollingResponse } from "../utils/rolling-response";
-import { v4 as uuidv4 } from "uuid";
 import * as AddressDAL from "../dal/address";
 import _ from "lodash";
+import { ObjectId } from "mongodb";
+import { getUser } from "../dal/user";
+import RollingError from "../utils/error";
 
 export async function getAllAddress(
   req: RollingTypes.Request,
 ): Promise<RollingResponse> {
   const { uid } = req.ctx.decodedToken;
+
+  const user = await getUser(uid, "getAllAddress");
+  if (!user) throw new RollingError(404, "User not found");
 
   const Addresses = await AddressDAL.getAllAddress(uid, "getAllAddress");
 
@@ -29,11 +34,16 @@ export async function createNewAddress(
     defaultAddress,
   } = req.body;
 
-  const addressId = uuidv4();
+  const user = await getUser(uid, "createNewAddress");
+
+  if (!user) {
+    throw new RollingError(404, "User not found");
+  }
+
   const address: RollingTypes.Address = {
     address1,
     address2,
-    addressId,
+    _id: new ObjectId(),
     name,
     landmark,
     state,
@@ -43,6 +53,7 @@ export async function createNewAddress(
     defaultAddress,
     createdAt: Date.now(),
     modifiedAt: Date.now(),
+    uid: user.uid,
   };
 
   await AddressDAL.addAddress(address, uid);
@@ -66,13 +77,15 @@ export async function updateAddress(
     defaultAddress,
   } = req.body;
 
+  const user = await getUser(uid, "deleteAddress");
+  if (!user) throw new RollingError(404, "user not found");
+
   const addressId = req.params["id"];
 
   const address = await AddressDAL.getAddressById(uid, addressId);
 
   const updatedAddress: RollingTypes.Address = {
     ...address,
-    addressId,
     name,
     address1,
     address2,
@@ -94,7 +107,15 @@ export async function deleteAddress(
   req: RollingTypes.Request,
 ): Promise<RollingResponse> {
   const { uid } = req.ctx.decodedToken;
+
+  const user = await getUser(uid, "deleteAddress");
+  if (!user) throw new RollingError(404, "user not found");
+
   const addressId = req.params["id"];
+
+  const address = await AddressDAL.getAddressById(uid, addressId);
+
+  if (!address) throw new RollingError(404, "address not found");
 
   await AddressDAL.deleteById(uid, addressId);
   return new RollingResponse("address deleted");
